changed live tile to notify when catching pokemons
