+++ PokemonGo-UWP/project.json
@@ -1,22 +1,23 @@
 ﻿{
-    "dependencies": {
-        "GeoExtensions": "1.0.0.3",
-        "HockeySDK.UWP": "4.1.3",
-        "Microsoft.NETCore.UniversalWindowsPlatform": "5.2.2",
-        "Newtonsoft.Json": "9.0.1",
-        "Octokit": "0.21.1",
-        "Template10": "1.1.12-preview-160712",
-        "UniversalMapControl": "0.3.36",
-        "WinRTXamlToolkit.Controls.Gauge.UWP": "2.0.0",
-        "Xam.Plugin.DeviceMotion": "1.1.2",
-        "XamlAnimatedGif": "1.1.4"
-    },
-    "frameworks": {
-        "uap10.0": {}
-    },
-    "runtimes": {
-        "win10-arm": {},
-        "win10-x86": {},
-        "win10-x64": {}
-    }
+  "dependencies": {
+    "GeoExtensions": "1.0.0.3",
+    "HockeySDK.UWP": "4.1.3",
+    "Microsoft.NETCore.UniversalWindowsPlatform": "5.2.2",
+    "Newtonsoft.Json": "9.0.1",
+    "NotificationsExtensions.Win10": "14332.0.2",
+    "Octokit": "0.21.1",
+    "Template10": "1.1.12-preview-160712",
+    "UniversalMapControl": "0.3.36",
+    "WinRTXamlToolkit.Controls.Gauge.UWP": "2.0.0",
+    "Xam.Plugin.DeviceMotion": "1.1.2",
+    "XamlAnimatedGif": "1.1.4"
+  },
+  "frameworks": {
+    "uap10.0": {}
+  },
+  "runtimes": {
+    "win10-arm": {},
+    "win10-x86": {},
+    "win10-x64": {}
+  }
 }
